// src/app/shared/interfaces/auxilium/shortcut-center/shortcut.interface.ts
export interface ShortcutDisplay {
    id?: number;
    title: string;
    description: string;
    category: string;
    itemId1: number;
    itemId2: number;
    itemId3: number;
    itemId4: number;
    itemId5: number;
    itemId6: number;
    itemId7: number;
    itemId8: number;
    itemId9: number;
    itemId10: number;
    itemCode1: string;
    itemCode2: string;
    itemCode3: string;
    itemCode4: string;
    itemCode5: string;
    itemCode6: string;
    itemCode7: string;
    itemCode8: string;
    itemCode9: string;
    itemCode10: string;
    billType1: string;
    billType2: string;
    billType3: string;
    billType4: string;
    billType5: string;
    billType6: string;
    billType7: string;
    billType8: string;
    billType9: string;
    billType10: string;
    quantity1: number;
    quantity2: number;
    quantity3: number;
    quantity4: number;
    quantity5: number;
    quantity6: number;
    quantity7: number;
    quantity8: number;
    quantity9: number;
    quantity10: number;
    addedBy: string;
    addedDate: string;
    modifiedBy: string;
    modifiedDate: string;
}

export type GetShortcutResponse = ShortcutDisplay[];